+++ src/modules/aborter/aborter.utils.ts
@@ -2,6 +2,9 @@ import { TimeoutErrorOptions } from '../../features/timeout/timeout-error';
 import { AbortError } from '../../features/abort-error';
 import { ErrorMessage } from './aborter.constants';
 
+/**
+ * @internal
+ */
 export const getAbortErrorByReason = (reason?: any): AbortError => {
   if (reason instanceof AbortError) {
     return reason;