+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Saborter Changelog
 
+## v2.2.0 (Match 31th, 2026)
+
+### New Features
+
+- Added the `abortSignalAny` utility function [#60](https://github.com/TENSIILE/saborter/pull/60)
+- Added integration functionality with the `@saborter/server` package [#60](https://github.com/TENSIILE/saborter/pull/60)
+- Improved JSdoc documentation for `Aborter` [#60](https://github.com/TENSIILE/saborter/pull/60)
+
+### Bug Fixes
+
+- Fixed a bug in the `debounce` and `setTimeoutAsync` utilities with overriding the `initiator` field in the `AbortError` error [#60](https://github.com/TENSIILE/saborter/pull/60)
+
 ## v2.1.0 (March 18th, 2026)
 
 ### New Features